📦 NEW: Create settings page and improve settings slice
