+++ src/components/header/index.jsx
@@ -1,2 +1,2 @@
+/* eslint-disable import/prefer-default-export */
 export { default as Header } from './Header'
-export { default as FilterItem } from './DarkmodeToggle'
