+++ src/components/settings/index.jsx
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as DarkmodeToggle } from './DarkmodeToggle'
